requirements.txt
solanapy
django
solders
requests
